optimizing the bundle size of /company
